% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top.R
\name{compute_cors}
\alias{compute_cors}
\title{computes unscaled correlation matrix}
\usage{
compute_cors(x, y)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{vector of observations}
}
\description{
This function computes an unscaled correlation matrix
}
