% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top.R
\name{construct_ints}
\alias{construct_ints}
\title{constructs selected interactions}
\usage{
construct_ints(x, top_ints, modifier = NULL)
}
\arguments{
\item{x}{matrix of predictors}

\item{top_ints}{k x 2 integer matrix of interaction terms}

\item{modifier}{effect modifier}
}
\description{
This function constructs interaction terms
}
\examples{

x <- matrix(rnorm(1000 * 500), ncol = 500)
y <- rnorm(1000)

cormat <- compute_cors(x, y)

int_idx <- filter_top(cormat, k = 50)

int_mat <- construct_ints(x, ints)
}
\seealso{
\code{\link[intscreen]{filter_top}}, \code{\link[intscreen]{compute_cors}}, and \code{\link[intscreen]{intscreen}}
}
