% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top.R
\name{cv_intscreen}
\alias{cv_intscreen}
\title{all computations for interaction screening}
\usage{
cv_intscreen(x, y, k = 100, nsplits = 10, train.frac = 0.75,
  fraction.in.thresh = 1)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{vector of observations}

\item{k}{integer number of top interactions to screen}
}
\description{
This function does all computations for interaction screening
}
\examples{

set.seed(1)
x <- matrix(rnorm(100 * 350), ncol = 350)
y <- rnorm(100) + x[,1] * x[,2] * 0.5 - x[,3] * x[,4] * 0.5

## require that each interaction be in the top 200 ints 95\% of the 15 splits
ints <- cv_intscreen(x, y, k = 200, nsplits = 15, fraction.in.thresh = 0.95)

ints$int_idx

## require that each interaction be in the top 50 ints 100\% of the 15 splits
ints <- cv_intscreen(x, y, k = 50, nsplits = 15, fraction.in.thresh = 1)

ints$int_idx
}
