% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top.R
\name{filter_top}
\alias{filter_top}
\title{filters top interactions using correlation matrix}
\usage{
filter_top(cormat, k = 50)
}
\arguments{
\item{cormat}{correlation matrix returned by \code{\link[intscreen]{compute_cors}}}

\item{k}{integer number of top interactions to screen}
}
\description{
This function filters top interactions using correlation matrix
}
\seealso{
\code{\link[intscreen]{compute_cors}}, \code{\link[intscreen]{construct_ints}}, and \code{\link[intscreen]{intscreen}}
}
