% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intnet.R
\name{intnet}
\alias{intnet}
\title{Cross validation interaction screening}
\usage{
intnet(x, y, which.cols = 1:ncol(x), k = 100, nsplits = 10,
  train.frac = 0.75, fraction.in.thresh = 1, verbose = FALSE,
  modifier = NULL, ...)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{vector of observations}

\item{which.cols}{integer vector indicating which columns of \code{x} to check for interactions}

\item{k}{integer number of top interactions to screen}

\item{nsplits}{integer number of cross validation splits to run. defaults to 10}

\item{train.frac}{fraction of data used for each split. defaults to 0.75}

\item{fraction.in.thresh}{fraction of times across the \code{nsplits} CV splits each
interaction is required in the top \code{k} interactions in order to be selected}

\item{modifier}{effect modifier}

\item{...}{arguments to be passed to \code{\link[glmnet]{glmnet}}}
}
\description{
This function implements CV interaction screening
}
\examples{

set.seed(1)
x <- matrix(rnorm(100 * 350), ncol = 350)
y <- rnorm(100) + x[,1] * x[,2] * 0.5 - x[,3] * x[,4] * 0.5


## require that each interaction be in the top 50 ints 100\% of the 15 splits
intmod <- intnet(x, y, k = 50, nsplits = 15, fraction.in.thresh = 1)

plot(intmod)
}
