% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top.R
\name{intscreen}
\alias{intscreen}
\title{all computations for interaction screening}
\usage{
intscreen(x, y, k = 10, modifier = NULL)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{vector of observations}

\item{k}{integer number of top interactions to screen}

\item{modifier}{effect modifier}
}
\description{
This function does all computations for interaction screening
}
\examples{

x <- matrix(rnorm(1000 * 500), ncol = 500)
y <- rnorm(1000)

ints <- intscreen(x, y, k = 12)

str(ints)
}
