% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top.R
\name{screen_ints}
\alias{screen_ints}
\title{all computations for interaction screening}
\usage{
screen_ints(x, y, k = 10)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{vector of observations}

\item{k}{integer number of top interactions to screen}
}
\description{
This function does all computations for interaction screening
}
\examples{

x <- matrix(rnorm(1000 * 500), ncol = 500)
y <- rnorm(1000)

cormat <- compute_cors(x, y)

int_idx <- filter_top(cormat, k = 50)

int_mat <- construct_ints(x, y, k = 12)

head(int_mat)
}
