% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_top.R
\name{cv_intscreen}
\alias{cv_intscreen}
\title{Cross validation interaction screening}
\usage{
cv_intscreen(x, y, nints = 100, nsplits = 10, train.frac = 0.75,
  fraction.in.thresh = 1, heredity = c("none", "weak", "strong"),
  verbose = FALSE, modifier = NULL)
}
\arguments{
\item{x}{matrix of predictors}

\item{y}{vector of observations}

\item{nints}{integer number of top interactions to screen}

\item{nsplits}{integer number of cross validation splits to run. defaults to 10}

\item{train.frac}{fraction of data used for each split. defaults to 0.75}

\item{fraction.in.thresh}{fraction of times across the \code{nsplits} CV splits each
interaction is required in the top \code{k} interactions in order to be selected}

\item{heredity}{either \code{"weak"}, \code{"strong"}, or \code{"none"}}

\item{verbose}{logical value, whether to print progress of the CV splitting}

\item{modifier}{effect modifier}
}
\description{
This function implements CV interaction screening
}
\examples{

set.seed(1)
x <- matrix(rnorm(100 * 350), ncol = 350)
y <- rnorm(100) + x[,1] * x[,2] * 0.5 - x[,3] * x[,4] * 0.5

## require that each interaction be in the top 200 ints 95\% of the 15 splits
ints <- cv_intscreen(x, y, nints = 200, nsplits = 15, fraction.in.thresh = 0.95)

ints$int_idx

## require that each interaction be in the top 50 ints 100\% of the 15 splits
ints <- cv_intscreen(x, y, nints = 50, nsplits = 15, fraction.in.thresh = 1)

ints$int_idx
}
